Make IlBuilder into more extensible class
In preparation for creating a JitBuilder extension for IlBuilder,
to facilitate the client API creation, the IlBuilder class needs
to become more like a fully extensible with an independent header.

Since compilertest provides an IlBuilder extension, it also needs
to be adjusted to account for the new structure.

Sigsned-off-by: Mark Stoodley <mstoodle@ca.ibm.com>
