+++ compiler/ilgen/OMRIlBuilder.cpp
@@ -84,10 +84,7 @@
 // objects have had injectIL() called.
 
 
-namespace OMR
-{
-
-IlBuilder::IlBuilder(TR::IlBuilder *source)
+OMR::IlBuilder::IlBuilder(TR::IlBuilder *source)
    : TR::IlInjector(source),
    _methodBuilder(source->_methodBuilder),
    _sequence(0),
@@ -102,13 +99,13 @@ IlBuilder::IlBuilder(TR::IlBuilder *source)
    }
 
 bool
-IlBuilder::TraceEnabled_log(){
+OMR::IlBuilder::TraceEnabled_log(){
     bool traceEnabled = _comp->getOption(TR_TraceILGen);
     return traceEnabled;
 }
 
 void
-IlBuilder::TraceIL_log(const char* s, ...){
+OMR::IlBuilder::TraceIL_log(const char* s, ...){
     va_list argp;
     va_start (argp, s);
     traceMsgVarArgs(_comp, s, argp);
@@ -116,14 +113,14 @@ IlBuilder::TraceIL_log(const char* s, ...){
 }
 
 void
-IlBuilder::initSequence()
+OMR::IlBuilder::initSequence()
    {
    _sequence = new (_comp->trMemory()->trHeapMemory()) List<SequenceEntry>(_comp->trMemory());
    _sequenceAppender = new (_comp->trMemory()->trHeapMemory()) ListAppender<SequenceEntry>(_sequence);
    }
 
 bool
-IlBuilder::injectIL()
+OMR::IlBuilder::injectIL()
    {
    TraceIL("Inside injectIL()\n");
    TraceIL("original entry %p\n", cfg()->getStart());
@@ -146,7 +143,7 @@ IlBuilder::injectIL()
    }
 
 void
-IlBuilder::setupForBuildIL()
+OMR::IlBuilder::setupForBuildIL()
    {
    initSequence();
    appendBlock(NULL, false);
@@ -155,7 +152,7 @@ IlBuilder::setupForBuildIL()
    }
 
 void
-IlBuilder::print(const char *title, bool recurse)
+OMR::IlBuilder::print(const char *title, bool recurse)
    {
    if (!TraceEnabled)
       return;
@@ -191,7 +188,7 @@ IlBuilder::print(const char *title, bool recurse)
    }
 
 void
-IlBuilder::printBlock(TR::Block *block)
+OMR::IlBuilder::printBlock(TR::Block *block)
    {
    if (!TraceEnabled)
       return;
@@ -207,21 +204,21 @@ IlBuilder::printBlock(TR::Block *block)
    }
 
 TR::SymbolReference *
-IlBuilder::lookupSymbol(const char *name)
+OMR::IlBuilder::lookupSymbol(const char *name)
    {
    TR_ASSERT(_methodBuilder, "cannot look up symbols in an IlBuilder that has no MethodBuilder");
    return _methodBuilder->lookupSymbol(name);
    }
 
 void
-IlBuilder::defineSymbol(const char *name, TR::SymbolReference *symRef)
+OMR::IlBuilder::defineSymbol(const char *name, TR::SymbolReference *symRef)
    {
    TR_ASSERT(_methodBuilder, "cannot define symbols in an IlBuilder that has no MethodBuilder");
    _methodBuilder->defineSymbol(name, symRef);
    }
 
 void
-IlBuilder::defineValue(const char *name, TR::IlType *type)
+OMR::IlBuilder::defineValue(const char *name, TR::IlType *type)
    {
    TR::DataType dt = type->getPrimitiveType();
    TR::SymbolReference *newSymRef = symRefTab()->createTemporary(methodSymbol(), dt);
@@ -230,7 +227,7 @@ IlBuilder::defineValue(const char *name, TR::IlType *type)
    }
 
 TR::IlValue *
-IlBuilder::newValue(TR::DataType dt, TR::Node *n)
+OMR::IlBuilder::newValue(TR::DataType dt, TR::Node *n)
    {
    // make sure TreeTop is well formed
    TR::Node *ttNode = n;
@@ -244,19 +241,19 @@ IlBuilder::newValue(TR::DataType dt, TR::Node *n)
    }
 
 TR::IlValue *
-IlBuilder::newValue(TR::IlType *dt, TR::Node *n)
+OMR::IlBuilder::newValue(TR::IlType *dt, TR::Node *n)
    {
    return newValue(dt->getPrimitiveType(), n);
    }
 
 TR::IlValue *
-IlBuilder::NewValue(TR::IlType *dt)
+OMR::IlBuilder::NewValue(TR::IlType *dt)
    {
    TR_ASSERT_FATAL(0, "should not create a value without a TR::Node");
    }
 
 TR::IlValue *
-IlBuilder::Copy(TR::IlValue *value)
+OMR::IlBuilder::Copy(TR::IlValue *value)
    {
    TR::DataType dt = value->getDataType();
    TR::SymbolReference *newSymRef = symRefTab()->createTemporary(_methodSymbol, dt);
@@ -276,7 +273,7 @@ IlBuilder::Copy(TR::IlValue *value)
    }
 
 TR::TreeTop *
-IlBuilder::getFirstTree()
+OMR::IlBuilder::getFirstTree()
    {
    TR_ASSERT(_blocks, "_blocks not created yet");
    TR_ASSERT(_blocks[0], "_blocks[0] not created yet");
@@ -284,7 +281,7 @@ IlBuilder::getFirstTree()
    }
 
 TR::TreeTop *
-IlBuilder::getLastTree()
+OMR::IlBuilder::getLastTree()
    {
    TR_ASSERT(_blocks, "_blocks not created yet");
    TR_ASSERT(_blocks[_numBlocks], "_blocks[0] not created yet");
@@ -292,7 +289,7 @@ IlBuilder::getLastTree()
    }
 
 uint32_t
-IlBuilder::countBlocks()
+OMR::IlBuilder::countBlocks()
    {
    // only count each block once
    if (_count > -1)
@@ -326,7 +323,7 @@ IlBuilder::countBlocks()
    }
 
 void
-IlBuilder::pullInBuilderTrees(TR::IlBuilder *builder,
+OMR::IlBuilder::pullInBuilderTrees(TR::IlBuilder *builder,
                               uint32_t *currentBlock,
                               TR::TreeTop **firstTree,
                               TR::TreeTop **newLastTree)
@@ -355,7 +352,7 @@ IlBuilder::pullInBuilderTrees(TR::IlBuilder *builder,
    }
 
 bool
-IlBuilder::connectTrees()
+OMR::IlBuilder::connectTrees()
    {
    // don't do this more than once per builder object
    if (_connectedTrees)
@@ -442,7 +439,7 @@ IlBuilder::connectTrees()
    }
 
 TR::Node *
-IlBuilder::zero(TR::DataType dt)
+OMR::IlBuilder::zero(TR::DataType dt)
    {
    switch (dt)
       {
@@ -456,19 +453,19 @@ IlBuilder::zero(TR::DataType dt)
    }
 
 TR::Node *
-IlBuilder::zero(TR::IlType *dt)
+OMR::IlBuilder::zero(TR::IlType *dt)
    {
    return zero(dt->getPrimitiveType());
    }
 
 TR::Node *
-IlBuilder::zeroNodeForValue(TR::IlValue *v)
+OMR::IlBuilder::zeroNodeForValue(TR::IlValue *v)
    {
    return zero(v->getDataType());
    }
 
 TR::IlValue *
-IlBuilder::zeroForValue(TR::IlValue *v)
+OMR::IlBuilder::zeroForValue(TR::IlValue *v)
    {
    TR::IlValue *returnValue = newValue(v->getDataType(), zeroNodeForValue(v));
    return returnValue;
@@ -476,7 +473,7 @@ IlBuilder::zeroForValue(TR::IlValue *v)
 
 
 TR::IlBuilder *
-IlBuilder::OrphanBuilder()
+OMR::IlBuilder::OrphanBuilder()
    {
    TR::IlBuilder *orphan = new (comp()->trHeapMemory()) TR::IlBuilder(_methodBuilder, _types);
    orphan->initialize(_details, _methodSymbol, _fe, _symRefTab);
@@ -485,7 +482,7 @@ IlBuilder::OrphanBuilder()
    }
 
 TR::Block *
-IlBuilder::emptyBlock()
+OMR::IlBuilder::emptyBlock()
    {
    TR::Block *empty = TR::Block::createEmptyBlock(NULL, _comp);
    cfg()->addNode(empty);
@@ -494,27 +491,27 @@ IlBuilder::emptyBlock()
 
 
 TR::IlBuilder *
-IlBuilder::createBuilderIfNeeded(TR::IlBuilder *builder)
+OMR::IlBuilder::createBuilderIfNeeded(TR::IlBuilder *builder)
    {
    if (builder == NULL)
       builder = OrphanBuilder();
    return builder;
    }
 
-IlBuilder::SequenceEntry *
-IlBuilder::blockEntry(TR::Block *block)
+OMR::IlBuilder::SequenceEntry *
+OMR::IlBuilder::blockEntry(TR::Block *block)
    {
    return new (_comp->trMemory()->trHeapMemory()) IlBuilder::SequenceEntry(block);
    }
 
-IlBuilder::SequenceEntry *
-IlBuilder::builderEntry(TR::IlBuilder *builder)
+OMR::IlBuilder::SequenceEntry *
+OMR::IlBuilder::builderEntry(TR::IlBuilder *builder)
    {
    return new (_comp->trMemory()->trHeapMemory()) IlBuilder::SequenceEntry(builder);
    }
 
 void
-IlBuilder::appendBlock(TR::Block *newBlock, bool addEdge)
+OMR::IlBuilder::appendBlock(TR::Block *newBlock, bool addEdge)
    {
    if (newBlock == NULL)
       {
@@ -534,7 +531,7 @@ IlBuilder::appendBlock(TR::Block *newBlock, bool addEdge)
    }
 
 void
-IlBuilder::AppendBuilder(TR::IlBuilder *builder)
+OMR::IlBuilder::AppendBuilder(TR::IlBuilder *builder)
    {
    TR_ASSERT(builder->_partOfSequence == false, "builder cannot be in two places");
 
@@ -555,19 +552,19 @@ IlBuilder::AppendBuilder(TR::IlBuilder *builder)
    }
 
 TR::Node *
-IlBuilder::loadValue(TR::IlValue *v)
+OMR::IlBuilder::loadValue(TR::IlValue *v)
    {
    return v->load(_currentBlock);
    }
 
 void
-IlBuilder::storeNode(TR::SymbolReference *symRef, TR::Node *v)
+OMR::IlBuilder::storeNode(TR::SymbolReference *symRef, TR::Node *v)
    {
    genTreeTop(TR::Node::createStore(symRef, v));
    }
 
 void
-IlBuilder::indirectStoreNode(TR::Node *addr, TR::Node *v)
+OMR::IlBuilder::indirectStoreNode(TR::Node *addr, TR::Node *v)
    {
    TR::DataType dt = v->getDataType();
    TR::SymbolReference *storeSymRef = symRefTab()->findOrCreateArrayShadowSymbolRef(dt, addr);
@@ -576,7 +573,7 @@ IlBuilder::indirectStoreNode(TR::Node *addr, TR::Node *v)
    }
 
 TR::IlValue *
-IlBuilder::indirectLoadNode(TR::IlType *dt, TR::Node *addr, bool isVectorLoad)
+OMR::IlBuilder::indirectLoadNode(TR::IlType *dt, TR::Node *addr, bool isVectorLoad)
    {
    TR_ASSERT(dt->isPointer(), "indirectLoadNode must apply to pointer type");
    TR::IlType * baseType = dt->baseType();
@@ -608,7 +605,7 @@ IlBuilder::indirectLoadNode(TR::IlType *dt, TR::Node *addr, bool isVectorLoad)
  * @param value IlValue that should be written to the local variable, which should be the same data type
  */
 void
-IlBuilder::Store(const char *varName, TR::IlValue *value)
+OMR::IlBuilder::Store(const char *varName, TR::IlValue *value)
    {
    if (!_methodBuilder->symbolDefined(varName))
       _methodBuilder->defineValue(varName, _types->PrimitiveType(value->getDataType()));
@@ -624,7 +621,7 @@ IlBuilder::Store(const char *varName, TR::IlValue *value)
  * @param value IlValue that should overwrite "dest"
  */
 void
-IlBuilder::StoreOver(TR::IlValue *dest, TR::IlValue *value)
+OMR::IlBuilder::StoreOver(TR::IlValue *dest, TR::IlValue *value)
    {
    dest->storeOver(value, _currentBlock);
    }
@@ -637,7 +634,7 @@ IlBuilder::StoreOver(TR::IlValue *dest, TR::IlValue *value)
  * @param value IlValue with the vector data that should be written to the local variable, and should have the same data type
  */
 void
-IlBuilder::VectorStore(const char *varName, TR::IlValue *value)
+OMR::IlBuilder::VectorStore(const char *varName, TR::IlValue *value)
    {
    TR::Node *valueNode = loadValue(value);
    TR::DataType dt = valueNode->getDataType();
@@ -661,7 +658,7 @@ IlBuilder::VectorStore(const char *varName, TR::IlValue *value)
  * @param value IlValue that should be written at "address"
  */
 void
-IlBuilder::StoreAt(TR::IlValue *address, TR::IlValue *value)
+OMR::IlBuilder::StoreAt(TR::IlValue *address, TR::IlValue *value)
    {
    TR_ASSERT(address->getDataType() == TR::Address, "StoreAt needs an address operand");
 
@@ -676,7 +673,7 @@ IlBuilder::StoreAt(TR::IlValue *address, TR::IlValue *value)
  * @param value IlValue with the vector data that should be written  at "address"
  */
 void
-IlBuilder::VectorStoreAt(TR::IlValue *address, TR::IlValue *value)
+OMR::IlBuilder::VectorStoreAt(TR::IlValue *address, TR::IlValue *value)
    {
    TR_ASSERT(address->getDataType() == TR::Address, "VectorStoreAt needs an address operand");
 
@@ -691,7 +688,7 @@ IlBuilder::VectorStoreAt(TR::IlValue *address, TR::IlValue *value)
    }
 
 TR::IlValue *
-IlBuilder::CreateLocalArray(int32_t numElements, TR::IlType *elementType)
+OMR::IlBuilder::CreateLocalArray(int32_t numElements, TR::IlType *elementType)
    {
    uint32_t size = numElements * elementType->getSize();
    TR::SymbolReference *localArraySymRef = symRefTab()->createLocalPrimArray(size,
@@ -712,7 +709,7 @@ IlBuilder::CreateLocalArray(int32_t numElements, TR::IlType *elementType)
    }
 
 TR::IlValue *
-IlBuilder::CreateLocalStruct(TR::IlType *structType)
+OMR::IlBuilder::CreateLocalStruct(TR::IlType *structType)
    {
    //similar to CreateLocalArray except writing a method in StructType to get the struct size
    uint32_t size = structType->getSize();
@@ -733,7 +730,7 @@ IlBuilder::CreateLocalStruct(TR::IlType *structType)
    }
 
 void
-IlBuilder::StoreIndirect(const char *type, const char *field, TR::IlValue *object, TR::IlValue *value)
+OMR::IlBuilder::StoreIndirect(const char *type, const char *field, TR::IlValue *object, TR::IlValue *value)
   {
   TR::SymbolReference *symRef = (TR::SymbolReference*)_types->FieldReference(type, field);
   TR::DataType fieldType = symRef->getSymbol()->getDataType();
@@ -743,7 +740,7 @@ IlBuilder::StoreIndirect(const char *type, const char *field, TR::IlValue *objec
   }
 
 TR::IlValue *
-IlBuilder::Load(const char *name)
+OMR::IlBuilder::Load(const char *name)
    {
    TR::SymbolReference *symRef = lookupSymbol(name);
    TR::Node *valueNode = TR::Node::createLoad(symRef);
@@ -752,7 +749,7 @@ IlBuilder::Load(const char *name)
    }
 
 TR::IlValue *
-IlBuilder::VectorLoad(const char *name)
+OMR::IlBuilder::VectorLoad(const char *name)
    {
    TR::SymbolReference *nameSymRef = lookupSymbol(name);
    TR::DataType returnType = nameSymRef->getSymbol()->getDataType();
@@ -766,7 +763,7 @@ IlBuilder::VectorLoad(const char *name)
    }
 
 TR::IlValue *
-IlBuilder::LoadIndirect(const char *type, const char *field, TR::IlValue *object)
+OMR::IlBuilder::LoadIndirect(const char *type, const char *field, TR::IlValue *object)
    {
    TR::SymbolReference *symRef = (TR::SymbolReference *)_types->FieldReference(type, field);
    TR::DataType fieldType = symRef->getSymbol()->getDataType();
@@ -776,7 +773,7 @@ IlBuilder::LoadIndirect(const char *type, const char *field, TR::IlValue *object
    }
 
 TR::IlValue *
-IlBuilder::LoadAt(TR::IlType *dt, TR::IlValue *address)
+OMR::IlBuilder::LoadAt(TR::IlType *dt, TR::IlValue *address)
    {
    TR_ASSERT(address->getDataType() == TR::Address, "LoadAt needs an address operand");
    TR::IlValue *returnValue = indirectLoadNode(dt, loadValue(address));
@@ -785,7 +782,7 @@ IlBuilder::LoadAt(TR::IlType *dt, TR::IlValue *address)
    }
 
 TR::IlValue *
-IlBuilder::VectorLoadAt(TR::IlType *dt, TR::IlValue *address)
+OMR::IlBuilder::VectorLoadAt(TR::IlType *dt, TR::IlValue *address)
    {
    TR_ASSERT(address->getDataType() == TR::Address, "LoadAt needs an address operand");
    TR::IlValue *returnValue = indirectLoadNode(dt, loadValue(address), true);
@@ -794,7 +791,7 @@ IlBuilder::VectorLoadAt(TR::IlType *dt, TR::IlValue *address)
    }
 
 TR::IlValue *
-IlBuilder::IndexAt(TR::IlType *dt, TR::IlValue *base, TR::IlValue *index)
+OMR::IlBuilder::IndexAt(TR::IlType *dt, TR::IlValue *base, TR::IlValue *index)
    {
    TR::IlType *elemType = dt->baseType();
    TR_ASSERT(base->getDataType() == TR::Address, "IndexAt must be called with a pointer base");
@@ -847,7 +844,7 @@ IlBuilder::IndexAt(TR::IlType *dt, TR::IlValue *base, TR::IlValue *index)
  * a pointer to the type of the field.
  */
 TR::IlValue *
-IlBuilder::StructFieldInstanceAddress(const char* structName, const char* fieldName, TR::IlValue* obj) {
+OMR::IlBuilder::StructFieldInstanceAddress(const char* structName, const char* fieldName, TR::IlValue* obj) {
    auto offset = typeDictionary()->OffsetOf(structName, fieldName);
    auto ptype = typeDictionary()->PointerTo(typeDictionary()->GetFieldType(structName, fieldName));
    TR::IlValue* offsetValue = NULL;
@@ -870,13 +867,13 @@ IlBuilder::StructFieldInstanceAddress(const char* structName, const char* fieldN
  * offset of all union fields is zero.
  */
 TR::IlValue *
-IlBuilder::UnionFieldInstanceAddress(const char* unionName, const char* fieldName, TR::IlValue* obj) {
+OMR::IlBuilder::UnionFieldInstanceAddress(const char* unionName, const char* fieldName, TR::IlValue* obj) {
    auto ptype = typeDictionary()->PointerTo(typeDictionary()->UnionFieldType(unionName, fieldName));
    return ConvertTo(ptype, obj);
 }
 
 TR::IlValue *
-IlBuilder::NullAddress()
+OMR::IlBuilder::NullAddress()
    {
    TR::IlValue *returnValue = newValue(Address, TR::Node::aconst(0));
    TraceIL("IlBuilder[ %p ]::%d is NullAddress\n", this, returnValue->getID());
@@ -884,7 +881,7 @@ IlBuilder::NullAddress()
    }
 
 TR::IlValue *
-IlBuilder::ConstInt8(int8_t value)
+OMR::IlBuilder::ConstInt8(int8_t value)
    {
    TR::IlValue *returnValue = newValue(Int8, TR::Node::bconst(value));
    TraceIL("IlBuilder[ %p ]::%d is ConstInt8 %d\n", this, returnValue->getID(), value);
@@ -892,7 +889,7 @@ IlBuilder::ConstInt8(int8_t value)
    }
 
 TR::IlValue *
-IlBuilder::ConstInt16(int16_t value)
+OMR::IlBuilder::ConstInt16(int16_t value)
    {
    TR::IlValue *returnValue = newValue(Int16, TR::Node::sconst(value));
    TraceIL("IlBuilder[ %p ]::%d is ConstInt16 %d\n", this, returnValue->getID(), value);
@@ -900,7 +897,7 @@ IlBuilder::ConstInt16(int16_t value)
    }
 
 TR::IlValue *
-IlBuilder::ConstInt32(int32_t value)
+OMR::IlBuilder::ConstInt32(int32_t value)
    {
    TR::IlValue *returnValue = newValue(Int32, TR::Node::iconst(value));
    TraceIL("IlBuilder[ %p ]::%d is ConstInt32 %d\n", this, returnValue->getID(), value);
@@ -908,7 +905,7 @@ IlBuilder::ConstInt32(int32_t value)
    }
 
 TR::IlValue *
-IlBuilder::ConstInt64(int64_t value)
+OMR::IlBuilder::ConstInt64(int64_t value)
    {
    TR::IlValue *returnValue = newValue(Int64, TR::Node::lconst(value));
    TraceIL("IlBuilder[ %p ]::%d is ConstInt64 %d\n", this, returnValue->getID(), value);
@@ -916,7 +913,7 @@ IlBuilder::ConstInt64(int64_t value)
    }
 
 TR::IlValue *
-IlBuilder::ConstFloat(float value)
+OMR::IlBuilder::ConstFloat(float value)
    {
    TR::Node *fconstNode = TR::Node::create(0, TR::fconst, 0);
    fconstNode->setFloat(value);
@@ -926,7 +923,7 @@ IlBuilder::ConstFloat(float value)
    }
 
 TR::IlValue *
-IlBuilder::ConstDouble(double value)
+OMR::IlBuilder::ConstDouble(double value)
    {
    TR::Node *dconstNode = TR::Node::create(0, TR::dconst, 0);
    dconstNode->setDouble(value);
@@ -936,7 +933,7 @@ IlBuilder::ConstDouble(double value)
    }
 
 TR::IlValue *
-IlBuilder::ConstString(const char * const value)
+OMR::IlBuilder::ConstString(const char * const value)
    {
    TR::IlValue *returnValue = newValue(Address, TR::Node::aconst((uintptrj_t)value));
    TraceIL("IlBuilder[ %p ]::%d is ConstString %p\n", this, returnValue->getID(), value);
@@ -944,7 +941,7 @@ IlBuilder::ConstString(const char * const value)
    }
 
 TR::IlValue *
-IlBuilder::ConstAddress(const void * const value)
+OMR::IlBuilder::ConstAddress(const void * const value)
    {
    TR::IlValue *returnValue = newValue(Address, TR::Node::aconst((uintptrj_t)value));
    TraceIL("IlBuilder[ %p ]::%d is ConstAddress %p\n", this, returnValue->getID(), value);
@@ -952,7 +949,7 @@ IlBuilder::ConstAddress(const void * const value)
    }
 
 TR::IlValue *
-IlBuilder::ConstInteger(TR::IlType *intType, int64_t value)
+OMR::IlBuilder::ConstInteger(TR::IlType *intType, int64_t value)
    {
    if      (intType == Int8)  return ConstInt8 ((int8_t)  value);
    else if (intType == Int16) return ConstInt16((int16_t) value);
@@ -964,7 +961,7 @@ IlBuilder::ConstInteger(TR::IlType *intType, int64_t value)
    }
 
 TR::IlValue *
-IlBuilder::ConvertTo(TR::IlType *t, TR::IlValue *v)
+OMR::IlBuilder::ConvertTo(TR::IlType *t, TR::IlValue *v)
    {
    TR::DataType typeFrom = v->getDataType();
    TR::DataType typeTo = t->getPrimitiveType();
@@ -979,7 +976,7 @@ IlBuilder::ConvertTo(TR::IlType *t, TR::IlValue *v)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedConvertTo(TR::IlType *t, TR::IlValue *v)
+OMR::IlBuilder::UnsignedConvertTo(TR::IlType *t, TR::IlValue *v)
    {
    TR::DataType typeFrom = v->getDataType();
    TR::DataType typeTo = t->getPrimitiveType();
@@ -994,7 +991,7 @@ IlBuilder::UnsignedConvertTo(TR::IlType *t, TR::IlValue *v)
    }
 
 TR::IlValue *
-IlBuilder::convertTo(TR::IlType *t, TR::IlValue *v, bool needUnsigned)
+OMR::IlBuilder::convertTo(TR::IlType *t, TR::IlValue *v, bool needUnsigned)
    {
    TR::DataType typeFrom = v->getDataType();
    TR::DataType typeTo = t->getPrimitiveType();
@@ -1008,7 +1005,7 @@ IlBuilder::convertTo(TR::IlType *t, TR::IlValue *v, bool needUnsigned)
    }
 
 TR::IlValue *
-IlBuilder::unaryOp(TR::ILOpCodes op, TR::IlValue *v)
+OMR::IlBuilder::unaryOp(TR::ILOpCodes op, TR::IlValue *v)
    {
    TR::Node *valueNode = loadValue(v);
    TR::Node *result = TR::Node::create(op, 1, valueNode);
@@ -1018,7 +1015,7 @@ IlBuilder::unaryOp(TR::ILOpCodes op, TR::IlValue *v)
    }
 
 void
-IlBuilder::doVectorConversions(TR::Node **leftPtr, TR::Node **rightPtr)
+OMR::IlBuilder::doVectorConversions(TR::Node **leftPtr, TR::Node **rightPtr)
    {
    TR::Node *    left  = *leftPtr;
    TR::DataType lType = left->getDataType();
@@ -1034,7 +1031,7 @@ IlBuilder::doVectorConversions(TR::Node **leftPtr, TR::Node **rightPtr)
    }
 
 TR::IlValue *
-IlBuilder::widenIntegerTo32Bits(TR::IlValue *v)
+OMR::IlBuilder::widenIntegerTo32Bits(TR::IlValue *v)
    {
    if (v->getDataType() == TR::Int8 || v->getDataType() == TR::Int16)
       return ConvertTo(Int32, v);
@@ -1043,7 +1040,7 @@ IlBuilder::widenIntegerTo32Bits(TR::IlValue *v)
    }
 
 TR::Node*
-IlBuilder::binaryOpNodeFromNodes(TR::ILOpCodes op,
+OMR::IlBuilder::binaryOpNodeFromNodes(TR::ILOpCodes op,
                                  TR::Node *leftNode,
                                  TR::Node *rightNode) 
    {
@@ -1066,7 +1063,7 @@ IlBuilder::binaryOpNodeFromNodes(TR::ILOpCodes op,
    }
 
 TR::IlValue *
-IlBuilder::binaryOpFromNodes(TR::ILOpCodes op,
+OMR::IlBuilder::binaryOpFromNodes(TR::ILOpCodes op,
                              TR::Node *leftNode,
                              TR::Node *rightNode) 
    {
@@ -1076,7 +1073,7 @@ IlBuilder::binaryOpFromNodes(TR::ILOpCodes op,
    } 
 
 TR::IlValue *
-IlBuilder::binaryOpFromOpMap(OpCodeMapper mapOp,
+OMR::IlBuilder::binaryOpFromOpMap(OpCodeMapper mapOp,
                              TR::IlValue *left,
                              TR::IlValue *right)
    {
@@ -1090,7 +1087,7 @@ IlBuilder::binaryOpFromOpMap(OpCodeMapper mapOp,
    }
 
 TR::IlValue *
-IlBuilder::binaryOpFromOpCode(TR::ILOpCodes op,
+OMR::IlBuilder::binaryOpFromOpCode(TR::ILOpCodes op,
                               TR::IlValue *left,
                               TR::IlValue *right)
    {
@@ -1101,7 +1098,7 @@ IlBuilder::binaryOpFromOpCode(TR::ILOpCodes op,
    }
 
 TR::IlValue *
-IlBuilder::compareOp(TR_ComparisonTypes ct,
+OMR::IlBuilder::compareOp(TR_ComparisonTypes ct,
                      bool needUnsigned,
                      TR::IlValue *left,
                      TR::IlValue *right)
@@ -1113,7 +1110,7 @@ IlBuilder::compareOp(TR_ComparisonTypes ct,
    }
 
 TR::IlValue *
-IlBuilder::NotEqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::NotEqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=compareOp(TR_cmpNE, false, left, right);
    TraceIL("IlBuilder[ %p ]::%d is NotEqualTo %d != %d?\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1121,14 +1118,14 @@ IlBuilder::NotEqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 void
-IlBuilder::Goto(TR::IlBuilder **dest)
+OMR::IlBuilder::Goto(TR::IlBuilder **dest)
    {
    *dest = createBuilderIfNeeded(*dest);
    Goto(*dest);
    }
 
 void
-IlBuilder::Goto(TR::IlBuilder *dest)
+OMR::IlBuilder::Goto(TR::IlBuilder *dest)
    {
    TR_ASSERT(dest != NULL, "This goto implementation requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::Goto %p\n", this, dest);
@@ -1137,7 +1134,7 @@ IlBuilder::Goto(TR::IlBuilder *dest)
    }
 
 void
-IlBuilder::Return()
+OMR::IlBuilder::Return()
    {
    TraceIL("IlBuilder[ %p ]::Return\n", this);
    TR::Node *returnNode = TR::Node::create(TR::ILOpCode::returnOpCode(TR::NoType));
@@ -1147,7 +1144,7 @@ IlBuilder::Return()
    }
 
 void
-IlBuilder::Return(TR::IlValue *value)
+OMR::IlBuilder::Return(TR::IlValue *value)
    {
    TraceIL("IlBuilder[ %p ]::Return %d\n", this, value->getID());
    TR::Node *returnNode = TR::Node::create(TR::ILOpCode::returnOpCode(value->getDataType()), 1, loadValue(value));
@@ -1157,7 +1154,7 @@ IlBuilder::Return(TR::IlValue *value)
    }
 
 TR::IlValue *
-IlBuilder::Sub(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Sub(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue = NULL;
    if (left->getDataType() == TR::Address)
@@ -1192,7 +1189,7 @@ static TR::ILOpCodes unsignedAddOpCode(TR::DataType type)
    }
 
 TR::IlValue *
-IlBuilder::Add(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Add(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue = NULL;
    if (left->getDataType() == TR::Address)
@@ -1225,7 +1222,7 @@ IlBuilder::Add(TR::IlValue *left, TR::IlValue *right)
  * blockAfterExceptionHandler:
  */
 void
-IlBuilder::appendExceptionHandler(TR::Block *blockThrowsException, TR::IlBuilder **handler, uint32_t catchType)
+OMR::IlBuilder::appendExceptionHandler(TR::Block *blockThrowsException, TR::IlBuilder **handler, uint32_t catchType)
    {
    //split block after overflowCHK, and add an goto to blockAfterExceptionHandler
    appendBlock();
@@ -1250,7 +1247,7 @@ IlBuilder::appendExceptionHandler(TR::Block *blockThrowsException, TR::IlBuilder
    }
 
 void
-IlBuilder::setHandlerInfo(uint32_t catchType)
+OMR::IlBuilder::setHandlerInfo(uint32_t catchType)
    {
    TR::Block *catchBlock = getEntry();
    catchBlock->setIsCold();
@@ -1258,7 +1255,7 @@ IlBuilder::setHandlerInfo(uint32_t catchType)
    }
 
 TR::Node*
-IlBuilder::genOverflowCHKTreeTop(TR::Node *operationNode, TR::ILOpCodes overflow)
+OMR::IlBuilder::genOverflowCHKTreeTop(TR::Node *operationNode, TR::ILOpCodes overflow)
    {
    TR::Node *overflowChkNode = TR::Node::createWithRoomForOneMore(overflow, 3, symRefTab()->findOrCreateOverflowCheckSymbolRef(_methodSymbol), operationNode, operationNode->getFirstChild(), operationNode->getSecondChild());
    overflowChkNode->setOverflowCheckOperation(operationNode->getOpCodeValue());
@@ -1267,7 +1264,7 @@ IlBuilder::genOverflowCHKTreeTop(TR::Node *operationNode, TR::ILOpCodes overflow
    }
 
 TR::IlValue *
-IlBuilder::genOperationWithOverflowCHK(TR::ILOpCodes op, TR::Node *leftNode, TR::Node *rightNode, TR::IlBuilder **handler, TR::ILOpCodes overflow)
+OMR::IlBuilder::genOperationWithOverflowCHK(TR::ILOpCodes op, TR::Node *leftNode, TR::Node *rightNode, TR::IlBuilder **handler, TR::ILOpCodes overflow)
    {
    /*
     * BB1:
@@ -1300,7 +1297,7 @@ IlBuilder::genOperationWithOverflowCHK(TR::ILOpCodes op, TR::Node *leftNode, TR:
 // This function takes 4 arguments and generate the addValue.
 // This function is called by AddWithOverflow and AddWithUnsignedOverflow.
 TR::ILOpCodes 
-IlBuilder::getOpCode(TR::IlValue *leftValue, TR::IlValue *rightValue)
+OMR::IlBuilder::getOpCode(TR::IlValue *leftValue, TR::IlValue *rightValue)
    {
    TR::ILOpCodes op;
    if (leftValue->getDataType() == TR::Address)
@@ -1317,7 +1314,7 @@ IlBuilder::getOpCode(TR::IlValue *leftValue, TR::IlValue *rightValue)
    }
 
 TR::IlValue *
-IlBuilder::AddWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::AddWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
    {
    TR::Node *leftNode = loadValue(left);
    TR::Node *rightNode = loadValue(right);
@@ -1328,7 +1325,7 @@ IlBuilder::AddWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlVal
    }
 
 TR::IlValue *
-IlBuilder::AddWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::AddWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
    {
    TR::Node *leftNode = loadValue(left);
    TR::Node *rightNode = loadValue(right);
@@ -1339,7 +1336,7 @@ IlBuilder::AddWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, T
    }
 
 TR::IlValue *
-IlBuilder::SubWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::SubWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
    {
    TR::Node *leftNode = loadValue(left);
    TR::Node *rightNode = loadValue(right);
@@ -1349,7 +1346,7 @@ IlBuilder::SubWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlVal
    }
 
 TR::IlValue *
-IlBuilder::SubWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::SubWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
    {
    TR::Node *leftNode = loadValue(left);
    TR::Node *rightNode = loadValue(right);
@@ -1359,7 +1356,7 @@ IlBuilder::SubWithUnsignedOverflow(TR::IlBuilder **handler, TR::IlValue *left, T
    }
 
 TR::IlValue *
-IlBuilder::MulWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::MulWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlValue *right)
    {
    TR::Node *leftNode = loadValue(left);
    TR::Node *rightNode = loadValue(right);
@@ -1369,7 +1366,7 @@ IlBuilder::MulWithOverflow(TR::IlBuilder **handler, TR::IlValue *left, TR::IlVal
    }
 
 TR::IlValue *
-IlBuilder::Mul(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Mul(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=binaryOpFromOpMap(TR::ILOpCode::multiplyOpCode, left, right);
    TraceIL("IlBuilder[ %p ]::%d is Mul %d * %d\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1377,7 +1374,7 @@ IlBuilder::Mul(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::Div(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Div(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=binaryOpFromOpMap(TR::ILOpCode::divideOpCode, left, right);
    TraceIL("IlBuilder[ %p ]::%d is Div %d / %d\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1385,7 +1382,7 @@ IlBuilder::Div(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::Rem(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Rem(TR::IlValue *left, TR::IlValue *right)
    {
    TR::DataType returnType = left->getDataType();
 
@@ -1406,7 +1403,7 @@ IlBuilder::Rem(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::And(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::And(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=binaryOpFromOpMap(TR::ILOpCode::andOpCode, left, right);
    TraceIL("IlBuilder[ %p ]::%d is And %d & %d\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1414,7 +1411,7 @@ IlBuilder::And(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::Or(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Or(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=binaryOpFromOpMap(TR::ILOpCode::orOpCode, left, right);
    TraceIL("IlBuilder[ %p ]::%d is Or %d | %d\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1422,7 +1419,7 @@ IlBuilder::Or(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::Xor(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::Xor(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=binaryOpFromOpMap(TR::ILOpCode::xorOpCode, left, right);
    TraceIL("IlBuilder[ %p ]::%d is Xor %d ^ %d\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1430,7 +1427,7 @@ IlBuilder::Xor(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::Node*
-IlBuilder::shiftOpNodeFromNodes(TR::ILOpCodes op,
+OMR::IlBuilder::shiftOpNodeFromNodes(TR::ILOpCodes op,
                                 TR::Node *leftNode,
                                 TR::Node *rightNode) 
    {
@@ -1442,7 +1439,7 @@ IlBuilder::shiftOpNodeFromNodes(TR::ILOpCodes op,
    }
 
 TR::IlValue *
-IlBuilder::shiftOpFromNodes(TR::ILOpCodes op,
+OMR::IlBuilder::shiftOpFromNodes(TR::ILOpCodes op,
                             TR::Node *leftNode,
                             TR::Node *rightNode) 
    {
@@ -1452,7 +1449,7 @@ IlBuilder::shiftOpFromNodes(TR::ILOpCodes op,
    } 
 
 TR::IlValue *
-IlBuilder::shiftOpFromOpMap(OpCodeMapper mapOp,
+OMR::IlBuilder::shiftOpFromOpMap(OpCodeMapper mapOp,
                             TR::IlValue *left,
                             TR::IlValue *right)
    {
@@ -1470,7 +1467,7 @@ IlBuilder::shiftOpFromOpMap(OpCodeMapper mapOp,
    }
 
 TR::IlValue *
-IlBuilder::ShiftL(TR::IlValue *v, TR::IlValue *amount)
+OMR::IlBuilder::ShiftL(TR::IlValue *v, TR::IlValue *amount)
    {
    TR::IlValue *returnValue=shiftOpFromOpMap(TR::ILOpCode::shiftLeftOpCode, v, amount);
    TraceIL("IlBuilder[ %p ]::%d is shr %d << %d\n", this, returnValue->getID(), v->getID(), amount->getID());
@@ -1478,7 +1475,7 @@ IlBuilder::ShiftL(TR::IlValue *v, TR::IlValue *amount)
    }
 
 TR::IlValue *
-IlBuilder::ShiftR(TR::IlValue *v, TR::IlValue *amount)
+OMR::IlBuilder::ShiftR(TR::IlValue *v, TR::IlValue *amount)
    {
    TR::IlValue *returnValue=shiftOpFromOpMap(TR::ILOpCode::shiftRightOpCode, v, amount);
    TraceIL("IlBuilder[ %p ]::%d is shr %d >> %d\n", this, returnValue->getID(), v->getID(), amount->getID());
@@ -1486,7 +1483,7 @@ IlBuilder::ShiftR(TR::IlValue *v, TR::IlValue *amount)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedShiftR(TR::IlValue *v, TR::IlValue *amount)
+OMR::IlBuilder::UnsignedShiftR(TR::IlValue *v, TR::IlValue *amount)
    {
    TR::IlValue *returnValue=shiftOpFromOpMap(TR::ILOpCode::unsignedShiftRightOpCode, v, amount);
    TraceIL("IlBuilder[ %p ]::%d is unsigned shr %d >> %d\n", this, returnValue->getID(), v->getID(), amount->getID());
@@ -1513,7 +1510,7 @@ IlBuilder::UnsignedShiftR(TR::IlValue *v, TR::IlValue *amount)
  * IfAnd(&inRange, &outOfRange, 2, cond1Builder, cond1, cond2Builder, cond2);
  */
 void
-IlBuilder::IfAnd(TR::IlBuilder **allTrueBuilder, TR::IlBuilder **anyFalseBuilder, int32_t numTerms, ...)
+OMR::IlBuilder::IfAnd(TR::IlBuilder **allTrueBuilder, TR::IlBuilder **anyFalseBuilder, int32_t numTerms, ...)
    {
    TR::IlBuilder *mergePoint = OrphanBuilder();
    *allTrueBuilder = createBuilderIfNeeded(*allTrueBuilder);
@@ -1565,7 +1562,7 @@ IlBuilder::IfAnd(TR::IlBuilder **allTrueBuilder, TR::IlBuilder **anyFalseBuilder
  * IfOr(&outOfRange, &inRange, 2, cond1Builder, cond1, cond2Builder, cond2);
  */
 void
-IlBuilder::IfOr(TR::IlBuilder **anyTrueBuilder, TR::IlBuilder **allFalseBuilder, int32_t numTerms, ...)
+OMR::IlBuilder::IfOr(TR::IlBuilder **anyTrueBuilder, TR::IlBuilder **allFalseBuilder, int32_t numTerms, ...)
    {
    TR::IlBuilder *mergePoint = OrphanBuilder();
    *anyTrueBuilder = createBuilderIfNeeded(*anyTrueBuilder);
@@ -1604,7 +1601,7 @@ IlBuilder::IfOr(TR::IlBuilder **anyTrueBuilder, TR::IlBuilder **allFalseBuilder,
    }
 
 TR::IlValue *
-IlBuilder::EqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::EqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    TR::IlValue *returnValue=compareOp(TR_cmpEQ, false, left, right);
    TraceIL("IlBuilder[ %p ]::%d is EqualTo %d == %d?\n", this, returnValue->getID(), left->getID(), right->getID());
@@ -1612,7 +1609,7 @@ IlBuilder::EqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 void
-IlBuilder::integerizeAddresses(TR::IlValue **leftPtr, TR::IlValue **rightPtr)
+OMR::IlBuilder::integerizeAddresses(TR::IlValue **leftPtr, TR::IlValue **rightPtr)
    {
    TR::IlValue *left = *leftPtr;
    if (left->getDataType() == TR::Address)
@@ -1624,7 +1621,7 @@ IlBuilder::integerizeAddresses(TR::IlValue **leftPtr, TR::IlValue **rightPtr)
    }
 
 TR::IlValue *
-IlBuilder::LessThan(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::LessThan(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpLT, false, left, right);
@@ -1633,7 +1630,7 @@ IlBuilder::LessThan(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedLessThan(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::UnsignedLessThan(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpLT, true, left, right);
@@ -1642,7 +1639,7 @@ IlBuilder::UnsignedLessThan(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::LessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::LessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpLE, false, left, right);
@@ -1651,7 +1648,7 @@ IlBuilder::LessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedLessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::UnsignedLessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpLE, true, left, right);
@@ -1660,7 +1657,7 @@ IlBuilder::UnsignedLessOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::GreaterThan(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::GreaterThan(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpGT, false, left, right);
@@ -1669,7 +1666,7 @@ IlBuilder::GreaterThan(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedGreaterThan(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::UnsignedGreaterThan(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpGT, true, left, right);
@@ -1678,7 +1675,7 @@ IlBuilder::UnsignedGreaterThan(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::GreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::GreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpGE, false, left, right);
@@ -1687,7 +1684,7 @@ IlBuilder::GreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue *
-IlBuilder::UnsignedGreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::UnsignedGreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    {
    integerizeAddresses(&left, &right);
    TR::IlValue *returnValue=compareOp(TR_cmpGE, true, left, right);
@@ -1696,7 +1693,7 @@ IlBuilder::UnsignedGreaterOrEqualTo(TR::IlValue *left, TR::IlValue *right)
    }
 
 TR::IlValue** 
-IlBuilder::processCallArgs(TR::Compilation *comp, int numArgs, va_list args)
+OMR::IlBuilder::processCallArgs(TR::Compilation *comp, int numArgs, va_list args)
    {
    TR::IlValue ** argValues = (TR::IlValue **) comp->trMemory()->allocateHeapMemory(numArgs * sizeof(TR::IlValue *));
    for (int32_t a=0;a < numArgs;a++)
@@ -1713,7 +1710,7 @@ IlBuilder::processCallArgs(TR::Compilation *comp, int numArgs, va_list args)
  *    The list is a computed address followed by the actual arguments
  */
 TR::IlValue *
-IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, ...)
+OMR::IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, ...)
    {
    TraceIL("IlBuilder[ %p ]::ComputedCall %s\n", this, functionName);
    va_list args;
@@ -1736,7 +1733,7 @@ IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, ...)
  *    the computed address followed by the actual arguments
  */
 TR::IlValue *
-IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, TR::IlValue **argValues)
+OMR::IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, TR::IlValue **argValues)
    {
    TraceIL("IlBuilder[ %p ]::ComputedCall %s\n", this, functionName);
    TR::ResolvedMethod *resolvedMethod = _methodBuilder->lookupFunction(functionName);
@@ -1749,7 +1746,7 @@ IlBuilder::ComputedCall(const char *functionName, int32_t numArgs, TR::IlValue *
    }
 
 TR::IlValue *
-IlBuilder::Call(const char *functionName, int32_t numArgs, ...)
+OMR::IlBuilder::Call(const char *functionName, int32_t numArgs, ...)
    {
    TraceIL("IlBuilder[ %p ]::Call %s\n", this, functionName);
    va_list args;
@@ -1766,7 +1763,7 @@ IlBuilder::Call(const char *functionName, int32_t numArgs, ...)
    }
 
 TR::IlValue *
-IlBuilder::Call(const char *functionName, int32_t numArgs, TR::IlValue ** argValues)
+OMR::IlBuilder::Call(const char *functionName, int32_t numArgs, TR::IlValue ** argValues)
    {
    TraceIL("IlBuilder[ %p ]::Call %s\n", this, functionName);
    TR::ResolvedMethod *resolvedMethod = _methodBuilder->lookupFunction(functionName);
@@ -1779,7 +1776,7 @@ IlBuilder::Call(const char *functionName, int32_t numArgs, TR::IlValue ** argVal
    }
 
 TR::IlValue *
-IlBuilder::genCall(TR::SymbolReference *methodSymRef, int32_t numArgs, TR::IlValue ** argValues, bool isDirectCall /* true by default*/)
+OMR::IlBuilder::genCall(TR::SymbolReference *methodSymRef, int32_t numArgs, TR::IlValue ** argValues, bool isDirectCall /* true by default*/)
    {
    TR::DataType returnType = methodSymRef->getSymbol()->castToMethodSymbol()->getMethod()->returnType();
    TR::Node *callNode = TR::Node::createWithSymRef(isDirectCall? TR::ILOpCode::getDirectCall(returnType): TR::ILOpCode::getIndirectCall(returnType), numArgs, methodSymRef);
@@ -1820,7 +1817,7 @@ IlBuilder::genCall(TR::SymbolReference *methodSymRef, int32_t numArgs, TR::IlVal
  *     The old value at the location \p baseAddress.
  */
 TR::IlValue *
-IlBuilder::AtomicAdd(TR::IlValue * baseAddress, TR::IlValue * value)
+OMR::IlBuilder::AtomicAdd(TR::IlValue * baseAddress, TR::IlValue * value)
    {
    TR_ASSERT(comp()->cg()->supportsAtomicAdd(), "this platform doesn't support AtomicAdd() yet");
    TR_ASSERT(baseAddress->getDataType() == TR::Address, "baseAddress must be TR::Address");
@@ -1913,7 +1910,7 @@ IlBuilder::AtomicAdd(TR::IlValue * baseAddress, TR::IlValue * value)
  *              |-->persistentFailurie
  */
 void
-IlBuilder::Transaction(TR::IlBuilder **persistentFailureBuilder, TR::IlBuilder **transientFailureBuilder, TR::IlBuilder **transactionBuilder)
+OMR::IlBuilder::Transaction(TR::IlBuilder **persistentFailureBuilder, TR::IlBuilder **transientFailureBuilder, TR::IlBuilder **transactionBuilder)
    {   
    //This assertion is to rule out platforms which don't have tstart evaluator yet. 
    TR_ASSERT(comp()->cg()->hasTMEvaluator(), "this platform doesn't support tstart or tfinish evaluator yet");   
@@ -1980,7 +1977,7 @@ IlBuilder::Transaction(TR::IlBuilder **persistentFailureBuilder, TR::IlBuilder *
  * Generate XABORT instruction to abort transaction
  */
 void
-IlBuilder::TransactionAbort()
+OMR::IlBuilder::TransactionAbort()
    {
    TraceIL("IlBuilder[ %p ]::transactionAbort", this);
    TR::Node *tAbortNode = TR::Node::create(TR::tabort, 0);
@@ -1989,14 +1986,14 @@ IlBuilder::TransactionAbort()
    }
 
 void
-IlBuilder::IfCmpNotEqualZero(TR::IlBuilder **target, TR::IlValue *condition)
+OMR::IlBuilder::IfCmpNotEqualZero(TR::IlBuilder **target, TR::IlValue *condition)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpNotEqualZero(*target, condition);
    }
 
 void
-IlBuilder::IfCmpNotEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
+OMR::IlBuilder::IfCmpNotEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
    {
    TR_ASSERT(target != NULL, "This IfCmpNotEqualZero requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpNotEqualZero %d? -> [ %p ] B%d\n", this, condition->getID(), target, target->getEntry()->getNumber());
@@ -2004,14 +2001,14 @@ IlBuilder::IfCmpNotEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
    }
 
 void
-IlBuilder::IfCmpNotEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpNotEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpNotEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpNotEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpNotEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpNotEqual requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpNotEqual %d == %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2019,14 +2016,14 @@ IlBuilder::IfCmpNotEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *
    }
 
 void
-IlBuilder::IfCmpEqualZero(TR::IlBuilder **target, TR::IlValue *condition)
+OMR::IlBuilder::IfCmpEqualZero(TR::IlBuilder **target, TR::IlValue *condition)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpEqualZero(*target, condition);
    }
 
 void
-IlBuilder::IfCmpEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
+OMR::IlBuilder::IfCmpEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
    {
    TR_ASSERT(target != NULL, "This IfCmpEqualZero requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpEqualZero %d == 0? -> [ %p ] B%d\n", this, condition->getID(), target, target->getEntry()->getNumber());
@@ -2034,14 +2031,14 @@ IlBuilder::IfCmpEqualZero(TR::IlBuilder *target, TR::IlValue *condition)
    }
 
 void
-IlBuilder::IfCmpEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpEqual requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpEqual %d == %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2049,14 +2046,14 @@ IlBuilder::IfCmpEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *rig
    }
 
 void
-IlBuilder::IfCmpLessThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpLessThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpLessThan(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpLessThan requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpLessThan %d < %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2064,14 +2061,14 @@ IlBuilder::IfCmpLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *
    }
 
 void
-IlBuilder::IfCmpUnsignedLessThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedLessThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpUnsignedLessThan(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpUnsignedLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpUnsignedLessThan requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpUnsignedLessThan %d < %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2079,14 +2076,14 @@ IlBuilder::IfCmpUnsignedLessThan(TR::IlBuilder *target, TR::IlValue *left, TR::I
    }
 
 void
-IlBuilder::IfCmpLessOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpLessOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpLessOrEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpLessOrEqual requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpLessOrEqual %d <= %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2094,14 +2091,14 @@ IlBuilder::IfCmpLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValu
    }
 
 void
-IlBuilder::IfCmpUnsignedLessOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedLessOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpUnsignedLessOrEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpUnsignedLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TR_ASSERT(target != NULL, "This IfCmpUnsignedLessOrEqual requires a non-NULL builder object");
    TraceIL("IlBuilder[ %p ]::IfCmpUnsignedLessOrEqual %d <= %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
@@ -2109,63 +2106,63 @@ IlBuilder::IfCmpUnsignedLessOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR
    }
 
 void
-IlBuilder::IfCmpGreaterThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpGreaterThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpGreaterThan(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpGreaterThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpGreaterThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TraceIL("IlBuilder[ %p ]::IfCmpGreaterThan %d > %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
    ifCmpCondition(TR_cmpGT, false, left, right, target->getEntry());
    }
 
 void
-IlBuilder::IfCmpUnsignedGreaterThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedGreaterThan(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpUnsignedGreaterThan(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpUnsignedGreaterThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedGreaterThan(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TraceIL("IlBuilder[ %p ]::IfCmpUnsignedGreaterThan %d > %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
    ifCmpCondition(TR_cmpGT, true, left, right, target->getEntry());
    }
 
 void
-IlBuilder::IfCmpGreaterOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpGreaterOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpGreaterOrEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpGreaterOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpGreaterOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TraceIL("IlBuilder[ %p ]::IfCmpGreaterOrEqual %d >= %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
    ifCmpCondition(TR_cmpGE, false, left, right, target->getEntry());
    }
 
 void
-IlBuilder::IfCmpUnsignedGreaterOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedGreaterOrEqual(TR::IlBuilder **target, TR::IlValue *left, TR::IlValue *right)
    {
    *target = createBuilderIfNeeded(*target);
    IfCmpUnsignedGreaterOrEqual(*target, left, right);
    }
 
 void
-IlBuilder::IfCmpUnsignedGreaterOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
+OMR::IlBuilder::IfCmpUnsignedGreaterOrEqual(TR::IlBuilder *target, TR::IlValue *left, TR::IlValue *right)
    {
    TraceIL("IlBuilder[ %p ]::IfCmpUnsignedGreaterOrEqual %d >= %d? -> [ %p ] B%d\n", this, left->getID(), right->getID(), target, target->getEntry()->getNumber());
    ifCmpCondition(TR_cmpGE, true, left, right, target->getEntry());
    }
 
 void
-IlBuilder::ifCmpCondition(TR_ComparisonTypes ct, bool isUnsignedCmp, TR::IlValue *left, TR::IlValue *right, TR::Block *target)
+OMR::IlBuilder::ifCmpCondition(TR_ComparisonTypes ct, bool isUnsignedCmp, TR::IlValue *left, TR::IlValue *right, TR::Block *target)
    {
    integerizeAddresses(&left, &right);
    TR::Node *leftNode = loadValue(left);
@@ -2179,19 +2176,19 @@ IlBuilder::ifCmpCondition(TR_ComparisonTypes ct, bool isUnsignedCmp, TR::IlValue
    }
 
 void
-IlBuilder::ifCmpNotEqualZero(TR::IlValue *condition, TR::Block *target)
+OMR::IlBuilder::ifCmpNotEqualZero(TR::IlValue *condition, TR::Block *target)
    {
    ifCmpCondition(TR_cmpNE, false, condition, zeroForValue(condition), target);
    }
 
 void
-IlBuilder::ifCmpEqualZero(TR::IlValue *condition, TR::Block *target)
+OMR::IlBuilder::ifCmpEqualZero(TR::IlValue *condition, TR::Block *target)
    {
    ifCmpCondition(TR_cmpEQ, false, condition, zeroForValue(condition), target);
    }
 
 void
-IlBuilder::appendGoto(TR::Block *destBlock)
+OMR::IlBuilder::appendGoto(TR::Block *destBlock)
    {
    gotoBlock(destBlock);
    appendBlock();
@@ -2207,7 +2204,7 @@ IlBuilder::appendGoto(TR::Block *destBlock)
  *     falls through to the merge point
  */
 void
-IlBuilder::IfThenElse(TR::IlBuilder **thenPath, TR::IlBuilder **elsePath, TR::IlValue *condition)
+OMR::IlBuilder::IfThenElse(TR::IlBuilder **thenPath, TR::IlBuilder **elsePath, TR::IlValue *condition)
    {
    TR_ASSERT(thenPath != NULL || elsePath != NULL, "IfThenElse needs at least one conditional path");
 
@@ -2274,7 +2271,7 @@ IlBuilder::IfThenElse(TR::IlBuilder **thenPath, TR::IlBuilder **elsePath, TR::Il
    }
 
 void
-IlBuilder::Switch(const char *selectionVar,
+OMR::IlBuilder::Switch(const char *selectionVar,
                   TR::IlBuilder **defaultBuilder,
                   uint32_t numCases,
                   int32_t *caseValues,
@@ -2336,7 +2333,7 @@ IlBuilder::Switch(const char *selectionVar,
    }
 
 void
-IlBuilder::Switch(const char *selectionVar,
+OMR::IlBuilder::Switch(const char *selectionVar,
                   TR::IlBuilder **defaultBuilder,
                   uint32_t numCases,
                   ...)
@@ -2376,7 +2373,7 @@ IlBuilder::Switch(const char *selectionVar,
 
 
 void
-IlBuilder::ForLoop(bool countsUp,
+OMR::IlBuilder::ForLoop(bool countsUp,
                    const char *indVar,
                    TR::IlBuilder **loopCode,
                    TR::IlBuilder **breakBuilder,
@@ -2440,7 +2437,7 @@ IlBuilder::ForLoop(bool countsUp,
    }
 
 void
-IlBuilder::DoWhileLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlBuilder **breakBuilder, TR::IlBuilder **continueBuilder)
+OMR::IlBuilder::DoWhileLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlBuilder **breakBuilder, TR::IlBuilder **continueBuilder)
    {
    methodSymbol()->setMayHaveLoops(true);
    TR_ASSERT(body != NULL, "doWhileLoop needs to have a body");
@@ -2478,7 +2475,7 @@ IlBuilder::DoWhileLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlB
    }
 
 void
-IlBuilder::WhileDoLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlBuilder **breakBuilder, TR::IlBuilder **continueBuilder)
+OMR::IlBuilder::WhileDoLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlBuilder **breakBuilder, TR::IlBuilder **continueBuilder)
    {
    methodSymbol()->setMayHaveLoops(true);
    TR_ASSERT(body != NULL, "WhileDo needs to have a body");
@@ -2510,5 +2507,3 @@ IlBuilder::WhileDoLoop(const char *whileCondition, TR::IlBuilder **body, TR::IlB
 
    AppendBuilder(done);
    }
-
-} // namespace OMR
